+++ Loon-0.5-Alpha/Loon-Neo/src/loon/action/sprite/effect/FadeDotEffect.java
@@ -230,12 +230,12 @@ public class FadeDotEffect extends LObject implements BaseEffect, ISprite {
 	}
 
 	@Override
-	public int getHeight() {
+	public float getHeight() {
 		return height;
 	}
 
 	@Override
-	public int getWidth() {
+	public float getWidth() {
 		return width;
 	}
